Restaurant fresh

+++ database/migrations/2020_11_12_111739_create_orders_table.php
@@ -125,7 +125,7 @@ class CreateOrdersTable extends Migration
             $table->string('order_comment')->nullable();
             $table->tinyInteger('order_feedback')->nullable();
             $table->string('feedback_comment')->nullable();
-            $table->tinyInteger('reason_id')->nullable();
+            $table->integer('reason_id')->nullable();
             $table->tinyInteger('user_type')->comment('1-Rider,2-Restaurent')->nullable();
             $table->tinyInteger('visibility')->default('0');
             $table->timestamp('deleted_at', 0)->nullable();
